requirements.txt
numpy
scipy
Pillow
scikit-learn
